Source code refactoring for Xampp 8.x
- Refactoring the source code
- Adding support for PHP 8.x

+++ xphp.bat
@@ -12,9 +12,10 @@ for /F "tokens=* USEBACKQ" %%v in (`where php`) do (
 set XPHP_APP_DIR=%~dp0
 if not "%XPHP_APP_DIR:~-2%"==":\" set XPHP_APP_DIR=%XPHP_APP_DIR:~0,-1%
 
+set XPHP_SRC_DIR=%XPHP_APP_DIR%\src
 set XPHP_TMP_DIR=%XPHP_APP_DIR%\tmp
-set XPHP_POWER_EXECUTOR=%XPHP_APP_DIR%\support\PowerExec.vbs
-set XPHP_PHP_CONTROLLER=%XPHP_APP_DIR%\xphp.php
+set XPHP_PHP_CONTROLLER=%XPHP_SRC_DIR%\xphp.php
+set XPHP_POWER_EXECUTOR=%XPHP_SRC_DIR%\power_exec.vbs
 
 if not exist "%XPHP_TMP_DIR%" mkdir "%XPHP_TMP_DIR%"
 goto startCommand
@@ -28,11 +29,21 @@ call %~fx0:clearEnvVars
 exit /B 1
 
 rem ---------------------------------------------
+:installed
+echo.
+echo Xampp PHP Switcher is already integrated into Xampp.
+echo No need to do it again.
+echo.
+pause>nul|set/p =Press any key to exit terminal...
+call :clearEnvVars
+exit /B
+
+rem ---------------------------------------------
 :installationFailed
 echo.
 echo Installation Xampp PHP Switcher failed.
 echo Please review the instructions carefully before installation.
-del /Q "%XPHP_APP_DIR%\settings.ini"
+if exist "%XPHP_APP_DIR%\settings.ini" del /Q "%XPHP_APP_DIR%\settings.ini"
 echo.
 pause>nul|set/p =Press any key to exit terminal...
 call :clearEnvVars
@@ -63,6 +74,7 @@ exit /B
 
 rem ---------------------------------------------
 :install
+if exist "%XPHP_APP_DIR%\settings.ini" goto installed
 FSUTIL dirty query %SystemDrive%>nul
 if %errorLevel% NEQ 0 (
     echo.
@@ -80,7 +92,14 @@ if %errorLevel% NEQ 0 (
 )
 
 php -n -d output_buffering=0 "%XPHP_PHP_CONTROLLER%" "install"
-if errorLevel 1 goto installationFailed
+
+if %errorLevel% EQU 1 goto installationFailed
+if %errorLevel% EQU 2 (
+    echo.
+    pause>nul|set/p =Press any key to exit terminal...
+    call :clearEnvVars
+    exit /B 2
+)
 
 echo|set/p =Moving directory of the current PHP build into the repository...
 for /F "tokens=* USEBACKQ" %%v in ("%XPHP_TMP_DIR%\.phpdir") do (set XPHP_PHP_DIR=%%v)
